+++ client/src/app.js
@@ -4,7 +4,7 @@ import NavigationBar from './components/shared/navigation/navigation-bar';
 import RouteComponent from './components/shared/navigation/navigation-route';
 import Home from './pages/home';
 import NoMatch from './pages/no-match';
-import info from './components/info/rotes';
+import info from './components/info/routes';
 
 class App extends Component {
   render() {
